Use mock func
