+++ pallets/parachain-staking/src/lib.rs
@@ -332,7 +332,7 @@ pub mod pallet {
 		/// Weight information for extrinsics in this pallet.
 		type WeightInfo: WeightInfo;
 
-		type CollatorDelegatorBlockRewardCalculator: CollatorDelegatorBlockRewardCalculator<Self>;
+		type BlockRewardCalculator: CollatorDelegatorBlockRewardCalculator<Self>;
 	}
 
 	#[pallet::error]
@@ -2678,13 +2678,13 @@ pub mod pallet {
 					.unwrap_or_else(Zero::zero);
 
 				let (now_read, now_write, now_reward) =
-					Self::collator_reward_per_block(&state, issue_number);
+					<T::BlockRewardCalculator as CollatorDelegatorBlockRewardCalculator<T>>::collator_reward_per_block(&state, issue_number);
 				Self::do_reward(&pot, &now_reward.owner, now_reward.amount);
 				reads = reads.saturating_add(now_read);
 				writes = writes.saturating_add(now_write);
 
 				let (now_read, now_write, now_rewards) =
-					Self::delegator_reward_per_block(&state, issue_number);
+					<T::BlockRewardCalculator as CollatorDelegatorBlockRewardCalculator<T>>::delegator_reward_per_block(&state, issue_number);
 				now_rewards.into_iter().for_each(|x| {
 					Self::do_reward(&pot, &x.owner, x.amount);
 				});
@@ -2820,73 +2820,6 @@ pub mod pallet {
 				Some(round.first + round.length),
 				// One read for the round info, blocknumber is read free
 				T::DbWeight::get().reads(1),
-			)
-		}
-	}
-
-	// [TODO] I'm thinking to extract a new pallet for that, otherwise, it's a little strange that
-	// we leave the reward here.
-	// Modifiy to return list
-	impl<T: Config> CollatorDelegatorBlockRewardCalculator<T> for Pallet<T> {
-		fn collator_reward_per_block(
-			stake: &Candidate<T::AccountId, BalanceOf<T>, T::MaxDelegatorsPerCollator>,
-			issue_number: BalanceOf<T>,
-		) -> (Weight, Weight, Reward<T::AccountId, BalanceOf<T>>) {
-			let min_delegator_stake = T::MinDelegatorStake::get();
-			let delegator_sum = (&stake.delegators)
-				.into_iter()
-				.filter(|x| x.amount >= min_delegator_stake)
-				.fold(T::CurrencyBalance::from(0u128), |acc, x| acc + x.amount);
-
-			let reward_rate_config = <RewardRateConfig<T>>::get();
-
-			if delegator_sum == T::CurrencyBalance::from(0u128) {
-				(
-					Weight::from_ref_time(1_u64),
-					Weight::from_ref_time(1_u64),
-					Reward { owner: stake.id.clone(), amount: issue_number },
-				)
-			} else {
-				let collator_reward = reward_rate_config.compute_collator_reward::<T>(issue_number);
-				(
-					Weight::from_ref_time(1_u64),
-					Weight::from_ref_time(1_u64),
-					Reward { owner: stake.id.clone(), amount: collator_reward },
-				)
-			}
-		}
-
-		fn delegator_reward_per_block(
-			stake: &Candidate<T::AccountId, BalanceOf<T>, T::MaxDelegatorsPerCollator>,
-			issue_number: BalanceOf<T>,
-		) -> (
-			Weight,
-			Weight,
-			BoundedVec<Reward<T::AccountId, BalanceOf<T>>, T::MaxDelegatorsPerCollator>,
-		) {
-			let min_delegator_stake = T::MinDelegatorStake::get();
-			let delegator_sum = (&stake.delegators)
-				.into_iter()
-				.filter(|x| x.amount >= min_delegator_stake)
-				.fold(T::CurrencyBalance::from(0u128), |acc, x| acc + x.amount);
-
-			let reward_rate_config = <RewardRateConfig<T>>::get();
-
-			let inner = (&stake.delegators)
-				.into_iter()
-				.filter(|x| x.amount >= min_delegator_stake)
-				.map(|x| {
-					let staking_rate = Perquintill::from_rational(x.amount, delegator_sum);
-					let delegator_reward = reward_rate_config
-						.compute_delegator_reward::<T>(issue_number, staking_rate);
-					Reward { owner: x.owner.clone(), amount: delegator_reward }
-				})
-				.collect::<Vec<Reward<T::AccountId, BalanceOf<T>>>>();
-
-			(
-				Weight::from_ref_time(1_u64 + 4_u64),
-				Weight::from_ref_time(inner.len() as u64),
-				inner.try_into().expect("Did not extend vec q.e.d."),
 			)
 		}
 	}
